+++ src/main/java/com/licyun/service/UserService.java
@@ -1,6 +1,8 @@
 package com.licyun.service;
 
 import com.licyun.model.User;
+import com.licyun.vo.MessageJsonBean;
+
 import java.util.List;
 
 
@@ -24,7 +26,7 @@ public interface UserService {
 
     void deleteUser(User user);
 
-    Integer findUserCount();
+    Long findUserCount();
 
     List<User> findAllUser();
